// types/metropole.ts
export interface Metropole {
  id: number
  name: string
  email: string
  cellPhone: string
  product: string
  interessePrincipal: string
  field01?: string
  field02?: string
  field03?: string
  field04?: string
  field05?: string
  field06?: string
  field07?: string
  field08?: string
  field09?: string
  field10?: string
  field11?: string
  field12?: string
  field13?: string
  field14?: string
  field15?: string
  field16?: string
  field17?: string
  field18?: string
  field19?: string
  field20?: string
  createdAt: string
  updatedAt: string
  tenantId?: {
    id: number
  }
}